Adding additional checks to stopOnlinePhysicalCopy
Added checks to stopOnlinePhysicalCopy to prevent errors
relating to 'copyOf' property missing.

Also added volume cleanup in other error cases to prevent
leftover volumes on the backend array.

Pinned flask to version 0.10.1

+++ hp3parclient/client.py
@@ -857,6 +857,7 @@ class HP3ParClient(object):
         task_id = None
         if task is None:
             # couldn't find the task
+            self.deleteVolume(name)
             msg = "Couldn't find the copy task for '%s'" % name
             raise exceptions.HTTPNotFound(error={'desc': msg})
         else:
@@ -867,6 +868,7 @@ class HP3ParClient(object):
             cmd = ['canceltask', '-f', task_id]
             self._run(cmd)
         else:
+            self.deleteVolume(name)
             msg = "Couldn't find the copy task for '%s'" % name
             raise exceptions.HTTPNotFound(error={'desc': msg})
 
@@ -882,11 +884,13 @@ class HP3ParClient(object):
         # now cleanup the dead snapshots
         vol = self.getVolume(name)
         if vol:
-            snap1 = self.getVolume(vol['copyOf'])
-            snap2 = self.getVolume(snap1['copyOf'])
+            if 'copyOf' in vol:
+                snap1 = self.getVolume(vol['copyOf'])
+                snap2 = self.getVolume(snap1['copyOf'])
             self.deleteVolume(name)
-            self.deleteVolume(snap1['name'])
-            self.deleteVolume(snap2['name'])
+            if 'copyOf' in vol:
+                self.deleteVolume(snap1['name'])
+                self.deleteVolume(snap2['name'])
 
     def getAllTasks(self):
         """Get the list of all Tasks
